src/requirements.txt
django >= 2.0
jinja2 >= 2.10
gunicorn >= 19.0
pillow >= 5.3.0
django-cors-headers
invoke ~= 1.2.0
django-environ==0.4.4
psycopg2-binary
toolz~=0.9.0
django-bulk-update~=2.2.0
django-manager-utils~=1.3.2
django-boogie~=0.13.2